Link multiple parties to multiple users #555
This patch creates a many to many relationship between party and nereid user.
This allows you to have any possibility of party-user relationship where a user
may switch current party to change parties. This also offers migration script
if the user had not installed the module and has existing users.