Update the migration instructions
Filling up the id also does the mistake of not incrementing the 
sequence of the table. When the user creates a record the new 
record id would conflict with the id. This change allows the 
migration script to just fill in nereid_user and party while 
the id is autofilled from the sequence